extract r2 bucket commands
